# src/ConsumerSimulator/DataModel/Product.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace DataModel {
    public class Product {
        public long Id { get; set; }
        public String Name { get; set; }
        public decimal Price { get; set; }
    }
}